Rename VBox extension to Hypervisor to make it generic for use by vboxmanage and a future libvirt module
